refactor: consolidate sample song logic into use-sample-songs hook; remove old sample-songs utility
